+++ lib/arel/nodes/select_configure.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Arel # :nodoc: all
   module Nodes
     class SelectConfigure < Unary