+++ lib/elasticsearch_record/gem_version.rb
@@ -9,7 +9,7 @@ module ElasticsearchRecord
   module VERSION
     MAJOR = 0
     MINOR = 1
-    TINY  = 0
+    TINY  = 1
     PRE   = nil
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")