+++ docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # ElasticsearchRecord - CHANGELOG
 
+## [0.1.1] - 2022-09-22
+* [add] msearch for klass & relation
+* [fix] Gemfile (remove duplicated requirements)
+* [fix] gemspec (dependencies & descriptions)
+* [fix] loading requirements, remove unused code
+* [fix] quoting
+* [fix] arel collector
+* [fix] arel visitor sort & assignment
+* [ref] results builder for multiple responses
+* [ref] log_subscriber to hide index but show 'query time'
+* [ref] model_schema attributes
+
 ## [0.1.0] - 2022-09-21
 * Initial commit
 * docs, version, structure